# internal/httpserver/handlers/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hostyt Proxy Gateway API",
    "version": "0.1.0",
...
  },
  "servers": [
    {
      "url": "https://proxy.host.yt",
      "description": "Production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key from `/admin/api-keys` or `/app/api-keys`."
      },
      "nodeAuthToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-node agent token. Also accepted as `?node_token=<token>`."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "not found" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/nodes/join": {
      "post": {
        "tags": ["Node Join"],
        "summary": "Register a Caddy node",
        "operationId": "nodeJoin",
        "description": "One-shot node registration. The `hpg_join_` token in the body IS the credential — no Bearer API key needed. Returns bootstrap payload consumed by the node-agent install script.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token"],
                "properties": {
                  "token": { "type": "string", "example": "hpg_join_xxxxxxxxxxxxxxxx" },
                  "public_hostname": { "type": "string", "example": "ams01.proxy.host.yt" },
                  "public_ip": { "type": "string", "format": "ipv4", "example": "185.10.20.30" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bootstrap payload",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "node_id": { "type": "integer", "format": "int64" },
                    "node_name": { "type": "string" },
                    "fingerprint": { "type": "string" },
                    "wireguard": { "type": "object", "description": "WireGuard interface config for the node" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid JSON or missing token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Token invalid or already used", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "429": { "description": "Rate limit exceeded", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}